backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "google>=3.0.0",
    "google-cloud-aiplatform>=1.83.0",
    "google-genai>=1.5.0",
    "jinja2>=3.1.6",
    "langchain-google-genai>=2.0.11",
    "orca>=1.8",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "psutil>=7.0.0",
    "uvicorn>=0.34.0",
]
